Fixed project structure (moved all form files into forms directory and removed extra form1.cs).
