+++ Forms/LoginForm.cs
@@ -1,0 +1,30 @@
+namespace MovieRentalProject
+{
+    public partial class LoginForm : Form
+    {
+        public LoginForm()
+        {
+            InitializeComponent();
+        }
+
+        private void UserBox_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void PassBox_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void LoginButton_Click(object sender, EventArgs e)
+        {
+            
+        }
+
+        private void Form1_Load(object sender, EventArgs e)
+        {
+            
+        }
+    }
+}
